+++ symbolic-debuginfo/src/function_builder.rs
@@ -63,19 +63,11 @@ pub struct FunctionBuilder<'s> {
     lines: Vec<LineInfo<'s>>,
     /// All variables found inside the function.
     variables: Vec<Variable<'s>>,
-    /// The maximum function parse depth we allow.
-    max_function_parse_depth: u32,
 }
 
 impl<'s> FunctionBuilder<'s> {
     /// Create a new builder for a given outer function.
-    pub fn new(
-        name: Name<'s>,
-        compilation_dir: &'s [u8],
-        address: u64,
-        size: u64,
-        max_function_parse_depth: u32,
-    ) -> Self {
+    pub fn new(name: Name<'s>, compilation_dir: &'s [u8], address: u64, size: u64) -> Self {
         Self {
             name,
             compilation_dir,
@@ -84,7 +76,6 @@ impl<'s> FunctionBuilder<'s> {
             inlinees: BinaryHeap::new(),
             lines: Vec::new(),
             variables: Vec::new(),
-            max_function_parse_depth,
         }
     }
 
@@ -94,7 +85,7 @@ impl<'s> FunctionBuilder<'s> {
     pub fn add_inlinee(&mut self, inlinee: FunctionBuilderInlinee<'s>) {
         // An inlinee that starts before the function is obviously bogus same for an inlinee that
         // has a depth deeper than the limit.
-        if inlinee.address < self.address || inlinee.depth > self.max_function_parse_depth {
+        if inlinee.address < self.address {
             return;
         }
 
@@ -145,7 +136,6 @@ impl<'s> FunctionBuilder<'s> {
             inlinees,
             variables,
             mut lines,
-            max_function_parse_depth: _,
         } = self;
 
         let inlinees = ensure_proper_nesting(inlinees)?;
@@ -524,7 +514,7 @@ mod tests {
     #[test]
     fn test_simple() {
         // 0x10 - 0x40: foo in foo.c on line 1
-        let mut builder = FunctionBuilder::new(Name::from("foo"), &[], 0x10, 0x30, 255);
+        let mut builder = FunctionBuilder::new(Name::from("foo"), &[], 0x10, 0x30);
         builder.add_leaf_line(0x10, Some(0x30), FileInfo::from_filename(b"foo.c"), 1);
         let func = builder.finish().unwrap();
 
@@ -537,7 +527,7 @@ mod tests {
         // 0x10 - 0x20: foo in foo.c on line 1
         // 0x20 - 0x40: bar in bar.c on line 1
         // - inlined into: foo in foo.c on line 2
-        let mut builder = FunctionBuilder::new(Name::from("foo"), &[], 0x10, 0x30, 255);
+        let mut builder = FunctionBuilder::new(Name::from("foo"), &[], 0x10, 0x30);
         builder.add_inlinee(FunctionBuilderInlinee {
             depth: 0,
             name: Name::from("bar"),
@@ -582,7 +572,7 @@ mod tests {
             }],
         };
 
-        let mut builder = FunctionBuilder::new(Name::from("outer"), &[], 0x10, 0x40, 255);
+        let mut builder = FunctionBuilder::new(Name::from("outer"), &[], 0x10, 0x40);
         builder.add_inlinee(FunctionBuilderInlinee {
             depth: 0,
             address: 0x20,
@@ -681,7 +671,7 @@ mod tests {
         //          |----|         |----| (parent.c line 1)
         //               |---------|      (child2.c line 1)
 
-        let mut builder = FunctionBuilder::new(Name::from("parent"), &[], 0x10, 0x40, 255);
+        let mut builder = FunctionBuilder::new(Name::from("parent"), &[], 0x10, 0x40);
         builder.add_inlinee(FunctionBuilderInlinee {
             depth: 0,
             name: Name::from("child1"),