+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Remove support for symcache version 7 and 8. ([#1033](https://github.com/getsentry/symbolic/pull/1033))
 - Adds variables to functions. ([#1025](https://github.com/getsentry/symbolic/pull/1025))
 - `SymCache` and `SourceLocation` no longer implement `PartialEq` and `Eq`. ([#1039](https://github.com/getsentry/symbolic/pull/1039))
+- Remove depth limits from function parsing. ([#1065](https://github.com/getsentry/symbolic/pull/1065))
 
 **Fixes**